Move `get_unresolved_peer_infos` and `set_peer_info` from `chia.util.config` to `chia.server.resolve_peer_info.py` so that `chia.util` no longer depends upon `chia.server`.
