+++ chia/cmds/configure.py
@@ -8,12 +8,12 @@ import yaml
 
 from chia.cmds.cmd_classes import ChiaCliContext
 from chia.server.outbound_message import NodeType
+from chia.server.resolve_peer_info import set_peer_info
 from chia.util.config import (
     initial_config_file,
     load_defaults_for_missing_services,
     lock_and_load_config,
     save_config,
-    set_peer_info,
     str2bool,
 )
